ID: 3509465 - Auto Archive Previous Medication on RePrescribe.
Some changes to review:
I moved the logging for "prescribe", until the script is written since
even though the action is the user clicking on ReRx the drug is not
represcribed until the script is written. So if a user accidentally
clicks on ReRx or changes their mind we are still logging that the
script as “represcribe”.  

Now when a drug is represcribed you will see in the log that the
medication and drugid was represcribed by a provider, the medication was
discontinued by the system and a new script was added by the provider.

The auto archive will happen when a user clicks on the ReRx link,
RePrescribe button and the RePrescribe Longterm Meds link from the
prescriptions module as well as the long term med feature from the
medication box on the encounter. But on all not until the save or save &
print buttons are selected.

For more description please review:

https://sourceforge.net/tracker/?func=detail&aid=3509465&group_id=66701&atid=515435

There are some images here as well to review.


Change-Id: I6bc785de16da2cfee5e95990d50d191fa8992e28