Bug Fix: ID 3588561 Bill To/Remit to Data Missing on 3rd Party Invoice
An extension to the previous fix submitted under the same change request ID, this
addition allows the user to edit the bill to info and override using the current
demographic contact info in the "Bill To" invoice reprint.
This addition still applies only if

useDemoClinicInfoOnInvoice=true

in oscar.properties (just like the previous commit under this change
request id).

Justification for change: In most cases it is desirable to use the current
demographic contact as the bill to, but in the case of an insurance
company paying the invoice, it is necessary to override using the
demographic contact info to reflect the insurance company contact.