Bug Fix: ID 3592471 PaymentReceived Report reports wrong OHIP Payment Date Range
If report is run (for example) between November 4th 2012 and December 4th
2012, the report picks up the November RA payments, instead of the
December RA payments, since MOH payment date in the ra_header table is
December 14th 2012, which doesn't fall into the specified reporting date
range.
Therefore, since the RA is received monthly, change the reporting query so
that it looks for MOH RA payments with the month of the specified end
date, rather than on the particular MOH payment date.