Bug Fix: ID 3594551 Multiple Same Day Bookings Warning Feature broken
Even if you attempt to add an appointment for a patient against a
different group schedule than the group schedule an existing appointment
is found on, on the same day, you will still get the "Multiple Same Day
Booking" warning message on the Add Appointment screen and the Add
Appointment buttons are hidden disallowing you to add the appointment.
This behaviour should only be true if the appointment was being added to
the same group schedule as the existing appointment.