ID: 3611774 ID: 3611963 Updated MySQL Connector
This commit leaves the database in its current (latin1) form but is part
of a series of commits that prepares for ISO-8 support

1) Pom updated to reference current mysql connector with stability and
speed fixes that can also handle UTF-8 lightly tested against one
servers 12_1 production data (with flag to protect against 3.0.11
0000-00-00 date glitch that we have depended on). Not pretty without
the flag with my data. Runs fine with the flag even if your data is
clean.

2) Dropped the mysql local repo as per Ted

Marc - rebased, fixed missing entries in pom.xml and oscar_mcmaster.properties

Change-Id: Id9ce847bce2042c0fc0a69304feb89b06313c5ae