ID 3613796 User Preference determines view of schedule
Currently, on login the default view of the schedule is "View All". This
change makes the default view of the schedule be a configurable user
preference, and works the same as Schedule Groups. If a user selects "View
All" or "Schedule View" in the top right-hand corner of the screen, that
becomes their new default. The user can also change this preference from
the Pref tab.