ID 3614598 A circumflex issues with Rx favorites When adding a new favorite, if the favorite name given has a mu symbol, an Fixes 2 issues:
A circumflex will precede each of these special characters in the Rx
favorite name when displayed on the favorites sidebar, or on the edit
favorites screen.

When clicking on an added favorite to use for a prescription, if the
favorite has a mu symbol in the drug name, the Name: field of the
prescription will contain a black diamond with a question mark where the
Mu symbol should appear.