3331 - Form Lab Req 2010 Letterhead selection shows non-doctor results
Changed the way that selections are handled. Given a selected physician who is
now disabled in the system, the selection will now turn red and a tooltip will appear
prompting the user to select a different letterhead. If the user chooses to save anyway
a message will appear prompting them to select a new letterhead.

Conflicts:
	src/main/java/org/oscarehr/PMmodule/dao/ProviderDao.java
	src/main/webapp/form/formlabreq10.jsp