#1100 Consult Print - Remove Referring Practitioner
The SFHC and MFP would like to remove the "Referring Practitioner" from
the consult print pdf. According to the clinics since this is populated
by the person editing the consult it is often in accurate and confusing
and adds to workload.

https://sourceforge.net/p/oscarmcmaster/feature-requests/1100/

Change-Id: I0913be5c01f497b448395190d52fb168e7cc73b8
(cherry picked from commit 26284478e9d886bf7e00afe5bd4d08d1afecd37b)